Safari-mac TAB behavior cleaned up and fixed RET-RET bug.
Tabbing in Safari-mac to another field was
  - moving the selection (good)
  - updating the input field (good)
  - moving focus to the next tab w/o ending the autocomplete (bad)
Now only cycles selection on TAB if active and lets the browser do
what it wants if not active.

Hitting RET-RET on an empty input field was:
  - activating autocomplete on RET (good)
  - throwing an error and remaining active on 2nd RET (bad)
Now it just closes on the second RET press.

Also changed the key calculation for keydown or keypressed events
to match browser key input quirks based on info at
  http://unixpapa.com/js/key.html