IE fixes for display of the suggestions and UP+DOWN arrow selection.
For some reason IE wraps the newly created wrapper for the suggestions
in a DIV.  Moving down to the wrapper node so it looks just like the
node all of the other browsers give us.

Keying off UP+DOWN on keyup because that is the only place IE exposes them.

Preventing TAB on both keyup and keydown during autocompleteMode so
we do not both select the next suggestion from our autocomplete list
and move focus to another element (i.e. suggestions for input A present
but focus is on input B).