Fixed autocomplete with scrollbar
IE[876] have problems when the ul has a scrollbar
and the user clicks there. The activate event
is thrown and the list disappears.

This patch tries to solve this.

Also passed jslint and google closure to check for
those typical mistakes.

I'm sorry for the conversion from spaces to tabs
:/